Update a few types

+++ src/psrc/common/pkd.c
@@ -15,7 +15,7 @@ int pkd_open(const char* p, bool nolock, struct pkd* d) {
     #if (PLATFLAGS & PLATFLAG_UNIXLIKE) || PLATFORM == PLAT_WIN32
         char* lockpath;
         if (!nolock) {
-            int lockpathlen = strlen(p);
+            size_t lockpathlen = strlen(p);
             lockpath = malloc(lockpathlen + 6);
             memcpy(lockpath, p, lockpathlen);
             lockpath[lockpathlen++] = '.';
